src/tool/custom_tools/git_agents/requirements.txt
requests==2.31.0
gitpython==3.1.40
pydantic==2.11.10
python-gitlab==4.0.0
PyGithub==1.59.1
playwright==1.40.0
pytest==7.4.3
pytest-asyncio==0.21.1
fastmcp==2.12.4
openai==1.85.0
fastapi==0.115.0
uvicorn==0.32.0